Move type to a less generic place
